changes in management, header include

+++ web/src/main/java/com/infoshareacademy/api/BookManagement.java
@@ -7,25 +7,23 @@ import javax.ws.rs.DELETE;
 import javax.ws.rs.Path;
 import javax.ws.rs.PathParam;
 import javax.ws.rs.core.Response;
-import java.util.Optional;
 
 @Path("/managemant")
 public class BookManagement {
 
     @Inject
     private BookService bookService;
+    private Enum REJECTED;
 
     @DELETE
     @Path("delete/{id}")
-    public Response deleteBook(@PathParam("id") String id){
+    public Response deleteBook(@PathParam("id") String id) {
 
+        if (id.matches("^[0-9]*$")) {
 
-        Optional.ofNullable(id).orElseThrow();
-        Long bookId = Long.parseLong(id);
-
-
-        return Response.ok().entity(bookService.deleteBook(bookId)).build();
-
-
+            Long bookId = Long.parseLong(id);
+            return Response.ok().entity(bookService.deleteBook(bookId)).build();
         }
+        return Response.ok().entity(REJECTED).build();
+    }
 }
